refactor: 상수 util 클래스로 제작
