+++ src/main/java/baseball/Application.java
@@ -1,5 +1,8 @@
 package baseball;
 
+import baseball.computer.Computer;
+import baseball.game.Game;
+import baseball.user.User;
 
 public class Application {
 	public static void main(String[] args) {
