+++ src/main/java/baseball/computer/Answer.java
@@ -1,7 +1,9 @@
-package baseball;
+package baseball.computer;
 
 import java.util.LinkedList;
 import java.util.List;
+
+import static baseball.util.ConstantUtil.*;
 
 import camp.nextstep.edu.missionutils.Randoms;
 
@@ -20,8 +22,8 @@ public class Answer {
 
 	public void makeAnswer() {
 		List<Integer> answer = new LinkedList<>();
-		while (answer.size() < 3) {
-			int number = Randoms.pickNumberInRange(1, 9);
+		while (answer.size() < ANSWER_SIZE) {
+			int number = Randoms.pickNumberInRange(ANSWER_MIN, ANSWER_MAX);
 			if (!answer.contains(number)) {
 				answer.add(number);
 			}
