+++ src/main/java/baseball/util/ConstantUtil.java
@@ -1,0 +1,10 @@
+package baseball.util;
+
+public class ConstantUtil {
+	public static final int RESTART = 1;
+	public static final int STOP = 2;
+
+	public final static int ANSWER_SIZE = 3;
+	public final static int ANSWER_MIN = 1;
+	public final static int ANSWER_MAX = 9;
+}
